Rename initial_import submodule
